fix(response): retain all the characters for the response message
Fix a bug in Mongoose which causes the last character in the response message to be replaced by null char if the message happens to be 8192 bytes long.

The Mongoose implementation calls vsnprintf which according to its documentation "if the resulting string would be longer than n-1 characters, the remaining characters are discarded and not stored, but counted for the value returned by the function". So if a message happens to be 8192 bytes long, the same length as the buffer size, then the 8192th char will be discarded, but will be counted.

While at mongoose.c#443, it is only reallocate longer buffer when the returned value from vsnprintf function call is greater than 8192, not when it is equal to 8192.

Closes #1