docs(readme): add a reminder that this is a repo used for ReadServer only
The reason why we need a separate fork for the ReadServer project is that we wanted to use an old version of mongoose plus a fix to a bug reported in issue #1.

The recent versions for mongoose completely change its interface which will cause major refactor if we use the latest version. While we have a separate fork and a new branch (readserver), we can easily set up submodules in the Read Server project and always points to the "readserver" branch.